pyproject.toml
[tool.poetry]
name = "amongusbot"
version = "0.1.0"
description = ""
authors = ["PederHA <peder.andresen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.4.1"
keyboard = "^0.13.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
